prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -10,7 +10,7 @@
 int main(void)
 {
 	long int x = 612852475143;
-	long int py;
+	int py;
 
 	for (py = 2; py < x; py++)
 	{
